# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koupii LMS API",
        "description": "API documentation for English course LMS",
        "contact": {
            "email": "support@koupii.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get all enrollments based on user role",
                "description": "Admin sees all enrollments, teacher sees only enrollments in their classes, student sees only their enrolled classes.",
                "operationId": "5e0b4253fceca57ce418fbc3f70f6d7b",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enrollments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "d40a0147-493f-4d70-bb4a-d7052e89e921"
                                            },
                                            "class_id": {
                                                "type": "string",
                                                "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                            },
                                            "student_id": {
                                                "type": "string",
                                                "example": "d2fb93ec-6043-4384-814d-0e48f36aed50"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "enrolled_at": {
                                                "type": "string",
                                                "example": "2025-07-17 08:55:35"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T08:55:35.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T08:55:35.000000Z"
                                            },
                                            "class": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                                    },
                                                    "teacher_id": {
                                                        "type": "string",
                                                        "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Theorytical Class"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "This is theory class"
                                                    },
                                                    "class_code": {
                                                        "type": "string",
                                                        "example": "lkoiun"
                                                    },
                                                    "cover_image": {
                                                        "type": "string",
                                                        "example": "cover.jpg"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "student": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "d2fb93ec-6043-4384-814d-0e48f36aed50"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Student User 2"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "student2@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "student"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/enrollments/create": {
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll student to class",
                "description": "Student enrolls to a class using class_id and class_code.",
                "operationId": "b047464eb13be51dd3d1870728618c77",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_id",
                                    "class_code",
                                    "status"
                                ],
                                "properties": {
                                    "class_id": {
                                        "type": "string",
                                        "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                    },
                                    "class_code": {
                                        "type": "string",
                                        "example": "lkoiun"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrolled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Enrolled successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "d40a0147-493f-4d70-bb4a-d7052e89e921"
                                                },
                                                "class_id": {
                                                    "type": "string",
                                                    "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid class code or validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/enrollments/{id}": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get enrollment details",
                "description": "Show enrollment details including class and student info.",
                "operationId": "307458d5057f1397dfd64576789a7ddc",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrollment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "d40a0147-493f-4d70-bb4a-d7052e89e921"
                                        },
                                        "class_id": {
                                            "type": "string",
                                            "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                        },
                                        "student_id": {
                                            "type": "string",
                                            "example": "d2fb93ec-6043-4384-814d-0e48f36aed50"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "enrolled_at": {
                                            "type": "string",
                                            "example": "2025-07-17 08:55:35"
                                        },
                                        "class": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Theorytical Class"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "student": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "d2fb93ec-6043-4384-814d-0e48f36aed50"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Student User 2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enrollment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}